Moved offline analysis to separate folder
Want to avoid a future where we have dozens of analyzers of different types, experiments, and scopes all living within one folder
