+++ ImageAnalysis/image_analysis/offline_analyzers/Undulator/ACaveMagCam3.py
@@ -5,10 +5,10 @@ from pathlib import Path
 
 if TYPE_CHECKING:
     from numpy.typing import NDArray
-    from ..types import Array2D
+    from image_analysis.types import Array2D
 
 import numpy as np
-from image_analysis.analyzers.basic_image_analysis import BasicImageAnalyzer
+from image_analysis.offline_analyzers.basic_image_analysis import BasicImageAnalyzer
 
 class ACaveMagCam3ImageAnalyzer(BasicImageAnalyzer):
